# perf(core): persist workspaces in parallel